refactor : Refactor code structure.
