--- project-6/capstone/project-dami-backend/src/models/DiscussionItem.ts
@@ -1,7 +1,0 @@
-export interface DiscussionItem {
-  title: string
-  shortDescription: string
-  createdAt: string
-  discussionId: string
-  userId: string
-}
