+++ project-6/capstone/project-dami-backend/src/models/discussion.ts
@@ -1,3 +1,11 @@
+export interface DiscussionItem {
+  title: string
+  shortDescription: string
+  createdAt: string
+  discussionId: string
+  userId: string
+}
+
 /**
  * Fields in a request to create a single DISCUSSION item.
  */
